+++ src/main/java/matteroverdrive/client/render/tileentity/TileEntityRendererPatternMonitor.java
@@ -22,7 +22,7 @@ public class TileEntityRendererPatternMonitor extends TileEntityRendererMonitor<
 
         int patternCount = 0;
 
-//        // tileEntity.getNetwork() is always null. Maybe because of the sidedness of the call? I.E. client vs. server.
+//         tileEntity.getNetwork() is always null. Maybe because of the sidedness of the call? I.E. client vs. server.
 //        if (tileEntity.getNetwork() != null && tileEntity.getConnectedDatabases() != null) {
 //            for (IMatterDatabase database: tileEntity.getConnectedDatabases()) {
 //                patternCount += database.getPatternStorageCount();