+++ src/main/java/matteroverdrive/gui/android/AndroidHudMinimap.java
@@ -86,7 +86,6 @@ public class AndroidHudMinimap extends AndroidHudElement {
         endMask();
 
         GlStateManager.enableTexture2D();
-        GlStateManager.enableDepth();
         GlStateManager.enableAlpha();
     }