+++ CHANGELOG.md
@@ -1,8 +1,9 @@
 MatterOverdrive - Changelog
 ===========================
 
-Unreleased
-----------
+0.7.5b2 (2022-01-29 21:01:19 +1300)
+-----------------------------------
+* Update CHANGELOG.md (infiniteblock)
 * Update RenderHandler.java (infiniteblock)
 * Fixes (infiniteblock)
 * Minimap Bugfix (infiniteblock)